requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
pandas==2.1.1
python-oracledb==1.4.1
python-dotenv==1.0.0
